+++ src/defaultProps.js
@@ -11,5 +11,6 @@
  * @param   {Object}    - component props
  * @returns {Function}
  */
-export default config => component => props =>
-	component({ ...config, ...props });
+export default config => Component => props => (
+	<Component {...config} {...props} />
+);