+++ src/shouldUpdate.js
@@ -5,13 +5,17 @@
  * @date   2017-01-06
  */
 
-import withLifecycle from './withLifecycle';
+// import withLifecycle from './withLifecycle';
 
 /**
  * @param   {Function}  - onComponentShouldUpdate - should update check function returning a boolean value
  * @param   {Function}  - component
- * @param   {Object}    - component props
  * @returns {Function}
  */
-export default onComponentShouldUpdate => component => props =>
-	withLifecycle({ onComponentShouldUpdate })(component)(props);
+export default onComponentShouldUpdate => Component => {
+	Component.defaultHooks = {
+		onComponentShouldUpdate,
+	};
+
+	return Component;
+};