+++ src/withHandlers.js
@@ -4,11 +4,11 @@
  * @date   2017-02-07
  */
 
-export default handlerCreators => component => props => {
+export default handlerCreators => Component => props => {
 	const composed = Object.keys(handlerCreators).reduce((acc, key) => {
 		acc[key] = handlerCreators[key](props);
 		return acc;
 	}, {});
 
-	return component({ ...props, ...composed });
+	return <Component {...props} {...composed} />;
 };