+++ tests/defaultProps.test.js
@@ -12,10 +12,10 @@ const defaultPropsValues = {
 	d: 4, // extends original
 };
 
-const component = props => props;
+const component = props => ({ props });
 
 test('should add additional props to components original props', () => {
-	const props = defaultProps(defaultPropsValues)(component)(originalProps);
+	const { props } = defaultProps(defaultPropsValues)(component)(originalProps);
 	expect(props).toEqual({
 		a: 1,
 		b: 2,