+++ tests/flattenProps.test.js
@@ -5,10 +5,10 @@ const originalProps = {
 	c: 3,
 };
 
-const component = props => props;
+const component = props => ({ props });
 
 test('should flatten prop `object`', () => {
-	const props = flattenProps('object')(component)(originalProps);
+	const { props } = flattenProps('object')(component)(originalProps);
 	expect(props).toEqual({
 		a: 1,
 		b: 2,