+++ tests/renameProp.test.js
@@ -6,10 +6,10 @@ const originalProps = {
 	c: 3,
 };
 
-const component = props => props;
+const component = props => ({ props });
 
 test('should rename prop `a` to `z`', () => {
-	const props = renameProp('a', 'z')(component)(originalProps);
+	const { props } = renameProp('a', 'z')(component)(originalProps);
 	expect(props).toEqual({
 		z: 1,
 		b: 2,