+++ tests/withHandlers.test.js
@@ -11,10 +11,10 @@ const withTestHandlers = {
 	getSum: props => () => props.a + props.b,
 };
 
-const component = props => props;
+const component = props => ({ props });
 
 test('should add handler creators and enrich them with props', () => {
-	const props = withHandlers(withTestHandlers)(component)(originalProps);
+	const { props } = withHandlers(withTestHandlers)(component)(originalProps);
 
 	expect(props.doTest).toEqual('successful? yup');
 	expect(props.getSum()).toBe(3);