+++ tests/withProps.test.js
@@ -12,10 +12,10 @@ const additionalProps = {
 	d: 4, // extends original
 };
 
-const component = props => props;
+const component = props => ({ props });
 
 test('should add additional props to components original props', () => {
-	const props = withProps(additionalProps)(component)(originalProps);
+	const { props } = withProps(additionalProps)(component)(originalProps);
 	expect(props).toEqual({
 		a: 10,
 		b: 20,